requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
argon2-cffi==23.1.0
Werkzeug==3.0.1
